added check receipt data section.

+++ soldi/interactive.py
@@ -9,10 +9,11 @@ from .table import *
 from .utils import *
 
 class Interactive(Config):
-    def __init__(self):
+    def __init__(self, debug=False):
         super().__init__()
         self.inputed_event_records = pd.DataFrame(index=[], columns=Event().cols)
         self.inputed_receipt_records = pd.DataFrame(index=[], columns=Receipt().cols)
+        self.writeout = False if debug else True
 
     def info(self):
         click.echo("=== Interactive Mode ===")
@@ -24,7 +25,7 @@ class Interactive(Config):
             event_record = event_prompt.start()
             self.inputed_event_records = self.inputed_event_records.append(event_record,
                                                                            ignore_index=True)
-
+            # Input receipt information.
             if event_record['_kind'] == "outgo":
                 msg = "\n" + soldi + \
                       " You can input receipt information of your inputed event."
@@ -35,29 +36,50 @@ class Interactive(Config):
                     self.inputed_receipt_records = pd.concat([self.inputed_receipt_records,
                                                               receipt_record],
                                                               ignore_index=True)
+                    # Check receipt data
+                    howmany_ser = self.inputed_receipt_records._howmany
+                    howmuch_ser = self.inputed_receipt_records._howmuch
+                    receipt_sum = (howmany_ser * howmuch_ser).sum()
+                    if self.inputed_event_records._howmuch.iloc[-1] == receipt_sum:
+                        click.echo(soldi + \
+                                   color.SUCCESS.format("Registered receipt data successfully"))
+                    elif self.inputed_event_records._howmuch.iloc[-1] > receipt_sum:
+                        warning_msg = soldi + color.WARNING.format(" The amount of inputted " \
+                                      "receipt is smaller than that of event data. Would you " \
+                                      "like to register with this content?")
+                        if not click.confirm(warning_msg):
+                            # もう一度レシートデータを登録するようにするべき
+                            break
+                    else:
+                        error_msg = soldi + color.ERROR.format(" The amount of inputted " \
+                                    "receipt is larger than that of event data.")
+                        click.echo(error_msg)
+                        break
 
             if not click.confirm("\n" + soldi + " Would you like to register other event data?"):
                 break
 
-        # Write out DataFrame
-        path_to_event = self.path_to_csv + "event.csv"
-        path_to_receipt = self.path_to_csv + "receipt.csv"
+        # Write out dataframes.
+        if self.writeout:
+            path_to_event = self.path_to_csv + "event.csv"
+            path_to_receipt = self.path_to_csv + "receipt.csv"
 
-        if path.isfile(path_to_event):
-            with open(path_to_event, 'a') as stream:
-                self.inputed_event_records.to_csv(stream,
-                                                  index=False, header=False)
-        else:
-            self.inputed_event_records.to_csv(path_to_event,
-                                              index=False, header=True)
-
-        if path.isfile(path_to_receipt):
-            with open(path_to_receipt, 'a') as stream:
-                self.inputed_receipt_records.to_csv(stream,
-                                                    index=False, header=False)
-        else:
-            self.inputed_receipt_records.to_csv(path_to_receipt,
-                                                index=False, header=True)
+            ## Write out a event dataframe.
+            if path.isfile(path_to_event):
+                with open(path_to_event, 'a') as stream:
+                    self.inputed_event_records.to_csv(stream,
+                                                      index=False, header=False)
+            else:
+                self.inputed_event_records.to_csv(path_to_event,
+                                                  index=False, header=True)
+            ## Write out a receipt dataframe.
+            if path.isfile(path_to_receipt):
+                with open(path_to_receipt, 'a') as stream:
+                    self.inputed_receipt_records.to_csv(stream,
+                                                        index=False, header=False)
+            else:
+                self.inputed_receipt_records.to_csv(path_to_receipt,
+                                                    index=False, header=True)
 
 
 class EventPrompt(object):
